requirements.txt
customtkinter
colorama
pyfiglet
Pillow
